Combine Donor with GKS
Due to the issue where Ganache is not letting me call 1 contract's
functions from another contract (even when it works fine in Remix
JavaScript VM and Rinkeby), combine the functions into 1 contract, so
that I can set up truffle tests.

Start with combining donor. Was able to register a donor ok as a test.

Signed-off-by: Nathalie C. Chan King Choy <nathalie.ckc.123@gmail.com>