Combine Charity, CRA, Auctioneer with GKS
Now all the functions are in 1 contract since both "truffle develop" and
ganache-cli development blockchains don't let me call the function in
another contract.

Did some basic testing in remix with ganache-cli to see that I could
register a donor, charity & auctioneer and do a donation that was
assigned by a charity to an auctioneer.

Signed-off-by: Nathalie C. Chan King Choy <nathalie.ckc.123@gmail.com>